Added dynamic normalization and separated custom slider component
